+++ obadiah/src/main/java/net/zionsoft/obadiah/model/analytics/Analytics.java
@@ -38,9 +38,7 @@ public class Analytics {
     public static final String DEEP_LINK_ACTION_OPENED = "opened";
 
     public static final String CATEGORY_NOTIFICATION = "notification";
-    public static final String NOTIFICATION_ACTION_ERROR = "error";
-    public static final String NOTIFICATION_ACTION_DEVICE_REGISTERED = "device_registered";
-    public static final String NOTIFICATION_ACTION_RECEIVED = "received";
+    public static final String NOTIFICATION_ACTION_REGISTERED = "device_registered";
     public static final String NOTIFICATION_ACTION_SHOWN = "shown";
     public static final String NOTIFICATION_ACTION_OPENED = "opened";
     public static final String NOTIFICATION_ACTION_DISMISSED = "dismissed";